Schema: Cart property type correction

+++ includes/api/cocart/v2/cart/class-cocart-cart-controller.php
@@ -2142,19 +2142,19 @@ class CoCart_Cart_V2_Controller extends CoCart_API_Controller {
 								'properties'  => array(
 									'value'        => array(
 										'description' => __( 'The quantity of the item in the cart.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'min_purchase' => array(
 										'description' => __( 'The minimum purchase amount required.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'max_purchase' => array(
 										'description' => __( 'The maximum purchase amount allowed. If -1 the item has an unlimited purchase amount.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
@@ -2168,25 +2168,25 @@ class CoCart_Cart_V2_Controller extends CoCart_API_Controller {
 								'properties'  => array(
 									'subtotal'     => array(
 										'description' => __( 'The subtotal of the item in the cart.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'subtotal_tax' => array(
 										'description' => __( 'The subtotal tax of the item in the cart.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'total'        => array(
 										'description' => __( 'The total of the item in the cart.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'total_tax'    => array(
 										'description' => __( 'The total tax of the item in the cart.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
@@ -2249,7 +2249,7 @@ class CoCart_Cart_V2_Controller extends CoCart_API_Controller {
 									),
 									'weight'       => array(
 										'description' => __( 'The weight of the item in the cart.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
@@ -2643,7 +2643,7 @@ class CoCart_Cart_V2_Controller extends CoCart_API_Controller {
 							),
 							'quantity'       => array(
 								'description' => __( 'The quantity of the item.', 'cart-rest-api-for-woocommerce' ),
-								'type'        => 'integer',
+								'type'        => 'float',
 								'context'     => array( 'view' ),
 								'readonly'    => true,
 							),
@@ -2655,25 +2655,25 @@ class CoCart_Cart_V2_Controller extends CoCart_API_Controller {
 								'properties'  => array(
 									'subtotal'     => array(
 										'description' => __( 'The subtotal of the item.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'subtotal_tax' => array(
 										'description' => __( 'The subtotal tax of the item.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'total'        => array(
 										'description' => __( 'The total of the item.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
 									'total_tax'    => array(
 										'description' => __( 'The total tax of the item.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
@@ -2736,7 +2736,7 @@ class CoCart_Cart_V2_Controller extends CoCart_API_Controller {
 									),
 									'weight'       => array(
 										'description' => __( 'The weight of the item.', 'cart-rest-api-for-woocommerce' ),
-										'type'        => 'integer',
+										'type'        => 'float',
 										'context'     => array( 'view' ),
 										'readonly'    => true,
 									),
